Move authentication hooks to SapphireTest
